connectivitycheckcontroller: use common code in library-go
